Added admin functionality to account info

+++ src/main/resources/static/js/views/UserPosts.js
@@ -1,15 +1,15 @@
 import createView from "../createView.js";
-import { getToken, isLoggedIn } from "../auth.js";
-import { getPosts, searchByCategory } from "../postFunctions.js";
+import {getToken} from "../auth.js";
+import {getAuthor, getPostCategories, getFormCategories} from "../postFunctions.js";
 
 export default function UserPosts(props) {
     // language=HTML
     return `
-        <header class="d-flex justify-content-end"
+        <header class=""
         </header>
         <main class="d-flex mx-2 h-100 justify-content-center">
             <div id="posts-container" class="d-flex flex-column justify-content-between mb-auto w-50">
-                
+
             </div>
             <form id="post-form" class="h-100 w-50">
                 <div class="text-center">
@@ -32,17 +32,41 @@ const getUserPosts = () => {
     fetch("http://localhost:8080/api/posts/getByUser", {method: "GET", headers: {Authorization: getToken()}})
         .then(results => results.json())
         .then(posts => {
-            console.log(posts)
-            $("#posts-container").html(getPosts(posts))
+            $("#posts-container").html(showUserPosts(posts))
         })
+}
+
+const showUserPosts = (posts) => {
+    //language=HTML
+    return posts.map(post =>
+        `
+            <div class="post-container card mx-1 mb-2 text-dark bg-transparent border-0">
+                <h3 id="title-${post.id}" class="card-title">${post.title}</h3>
+                <p id="content-${post.id}" class="card-body">${post.content}</p>
+                <div class="d-flex justify-content-end">${getPostCategories(post.categories)}</div>
+                <div class="card-footer d-flex justify-content-between bg-transparent mb-5">
+                    <div>
+                        <button data-id="${post.id}" class="edit-btn btn btn-sm btn-primary"
+                        ">Edit Post</button>
+                        <button data-id="${post.id}" class="del-btn btn btn-sm btn-primary"
+                        ">Delete Post</button>
+                    </div>
+                    <div>
+                        <p>${getAuthor(post.author)}</p>
+                    </div>
+                </div>
+            </div>
+        `).join('')
 }
 
 const submit = _ => {
     let id = null;
-    $(".edit-btn").click((e) => {
-        id = e.target.getAttribute("data-id")
-        $("#title").val($(`#title-${id}`).html())
-        $("#content").val($(`#content-${id}`).html())
+    $("main").click((e) => {
+        if (e.target.classList.contains("edit-btn")) {
+            id = e.target.getAttribute("data-id")
+            $("#title").val($(`#title-${id}`).html())
+            $("#content").val($(`#content-${id}`).html())
+        }
     });
     $('#submit-btn').click(function () {
         id === null ? createPost() : editPost(id)
@@ -68,10 +92,10 @@ const createPost = _ => {
     fetch(`http://localhost:8080/api/posts?categories=${categories}`, postRequest)
         .then(res => {
             console.log(res.status)
-            createView("/posts")
+            createView("/userPosts")
         }).catch(error => {
         console.log(error);
-        createView("/posts");
+        createView("/userPosts");
     });
 }
 
@@ -93,22 +117,24 @@ const editPost = id => {
 
     fetch(`http://localhost:8080/api/posts/${id}?categories=${categories}`, editRequest)
         .then(_ => {
-            createView("/posts")
+            createView("/userPosts")
         }).catch(_ => {
-        createView("/posts");
+        createView("/userPosts");
     });
 }
 
 const deletePost = _ => {
-    $(".del-btn").click((e) => {
-        const index = parseInt((e.target.getAttribute("data-id")))
-        fetch(`http://localhost:8080/api/posts/${index}`, {method: "DELETE", headers: {Authorization: getToken()}})
-            .then(res => {
-                createView("/posts")
-            }).catch(error => {
-            console.log(error);
-            createView("/posts");
-        });
+    $("main").click((e) => {
+        if (e.target.classList.contains("del-btn")) {
+            const index = parseInt((e.target.getAttribute("data-id")))
+            fetch(`http://localhost:8080/api/posts/${index}`, {method: "DELETE", headers: {Authorization: getToken()}})
+                .then(res => {
+                    createView("/userPosts")
+                }).catch(error => {
+                console.log(error);
+                createView("/userPosts");
+            });
+        }
     });
 }
 
@@ -120,10 +146,9 @@ const clearForm = _ => {
 }
 
 export function UserPostsEvent() {
-    showButtons();
     submit();
     deletePost();
     clearForm();
-    searchByCategory();
     getUserPosts();
+    getFormCategories();
 }
